+++ web-local/src/routes/(application)/welcome/+page.svelte
